refactor(style): move styles to gatsby-browser.js
