+++ gatsby-browser.js
@@ -1,0 +1,4 @@
+'use strict'
+
+require('./src/assets/scss/init.scss')
+require('./static/css/prismjs-theme.css')
