+++ src/assets/scss/init.scss
@@ -1,0 +1,28 @@
+@charset "UTF-8";
+
+/*
+figure {
+  display: block;
+  width: 100%;
+  height: auto;
+}
+
+figcaption {
+  font-style: italic;
+  margin-bottom: 0;
+  text-align: center;
+}
+
+figure.float-left,
+figure.float-right {
+  max-width: 310px;
+}
+
+.float-right {
+  float: right;
+}
+
+.float-left {
+  float: left;
+}
+*/
