+++ src/templates/blog-post-template.js
@@ -1,13 +1,11 @@
 import React from 'react'
-import Helmet from 'react-helmet'
 import { graphql, Link } from 'gatsby'
-import get from 'lodash/get'
 
 import Bio from '../components/Bio'
 import Layout from '../components/Layout'
 import { rhythm, scale } from '../utils/typography'
 import { Content, SectionTitle } from '../components/Utils'
-import '../utils/prismjs-theme.css'
+import '../../static/css/prismjs-theme.css'
 import { useSiteMetadata } from '../hooks'
 
 const BlogPostTemplate = ({ data, pageContext }) => {
